% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addWeights.r
\docType{methods}
\name{addWeights}
\alias{addWeights}
\alias{addWeights<-,dataObj,list-method}
\alias{addWeights<-,simPopObj,list-method}
\title{Methods for function \code{addWeights}}
\usage{
addWeights(object) <- value

\S4method{addWeights}{dataObj,list}(object) <- value

\S4method{addWeights}{simPopObj,list}(object) <- value
}
\description{
allows to modify sampling weights of an \code{\linkS4class{dataObj}} or
\code{\linkS4class{simPopObj}}-object. As input the output of
\code{\link{calibSample}} must be used.
}
\section{Methods}{
 \describe{
\item{list(signature(object="dataObj", value="list"))}{ \code{addWeight(inp) <- x} replaces the variable in slot
"data" containing sampling weights (specified by slot "weight" of input
\code{inp} with \code{x}. \code{x} must be a list that was computed using
\code{\link{calibWeights}}. }
\item{list(signature(object="simPopObj", value="list"))}{ \code{addWeight(inp) <- x} updates slot "sample" of the
provided object of class \code{\linkS4class{simPopObj}}. It replaces the
variable in slot "data" that contains sampling weights (specified by slot
"weight") of this slot with \code{x}. \code{x} must be a list that was
computed using \code{\link{calibWeights}}. }}
}
\examples{
data(eusilcS)
data(totalsRG)
inp <- specifyInput(data=eusilcS, hhid="db030", hhsize="hsize", strata="db040", weight="db090")
\dontrun{
## approx. 20 seconds ...
addWeights(inp) <- calibSample(inp, totalsRG)
}
}
\keyword{methods}

