% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{combine_factors}
\alias{combine_factors}
\title{Combine multiple factorvariables to a single one}
\usage{
combine_factors(dat, factor_columns, asfactor = TRUE)
}
\arguments{
\item{dat}{A \code{data.frame} object that holds several factor variables.}

\item{factor_columns}{A \code{character} naming all factors to be combined}

\item{asfactor}{A \code{logical} of length one, specifying whether the return value should be a \code{factor} (default)
or an \code{intger}.}
}
\description{
In order to apply \link{ipu_step} to multidimensional contigency-tables, this function is used to reduce
several factor variables to a single factor variable. The number of levels for the result is the
product of the number of levels of the input factors.
}
\examples{

################## create data ##################
factors <- c("time", "sex", "smoker", "day")
tips <- reshape2::tips[factors]

################ combine factors ################
cf <- combine_factors(tips, names(tips))
cbind(tips, cf)[sample(nrow(tips), 10),]

############# high dimensional ipu ##############
con <- xtabs(~., tips)
weight <- rnorm(nrow(tips)) + 5
adjusted_weight <- ipu_step(weight, cf, con)

#################### check ######################
con2 <- xtabs(adjusted_weight ~ ., data = tips)
sum((con - con2)^2)

}
