% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctHeap.R
\name{correctHeap}
\alias{correctHeap}
\title{Correct 5-year age heaping}
\usage{
correctHeap(x, method = "lnorm")
}
\arguments{
\item{x}{numeric vector}

\item{method}{default is log-normal (\dQuote{lnorm}). Method \dQuote{norm} and \dQuote{unif} are
also available.}
}
\description{
Correct fo age heaping using truncated (log-)normal distributions
}
\details{
for method lnorm, a truncated log-normal is fit to the whole age distribution. Then
for each age heaps (0,5,10,15, ...) random numbers of truncated log-normal is drawn in the 
interval +- 2 of the age heap for a certain ratio of values at an age heap. 
This ratio is taken from the mean of +- 1 of age regarding to the value at the age heap.
}
\examples{
## test with toy data (to see which methods works better)
## meanlog and sdlog estimated from real data.
y <- rlnorm(10000, meanlog = 2.466869, sdlog = 1.652772)
y <- round(y[y < 90])
plot(table(y))
w <- y \%in\% seq(0,100, by=5)
z <- c(y, y[w])
z <- sample(z)
## same size as y
z <- z[1:length(y)]
plot(table(z))
a1 <- correctHeap(z)
a2 <- correctHeap(z, method = "norm")
a3 <- correctHeap(z, method = "unif")
plot(table(a1))
plot(table(a2))
plot(table(a3))
sum(abs(table(a1) - table(y)))
sum(abs(table(a2) - table(y)))
sum(abs(table(a3) - table(y)))
}
\author{
Matthias Templ
}

