% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{eusilc13puf}
\alias{eusilc13puf}
\title{Synthetic EU-SILC 2013 survey data}
\format{A data frame with 13513 observations on the following 62 variables.
\describe{
\item{db030}{integer; the household ID.}
\item{hsize}{integer; the number of persons in the household.}
\item{db040}{factor; the federal state in which the household is
located (levels \code{Burgenland}, \code{Carinthia}, \code{Lower Austria},
\code{Salzburg}, \code{Styria}, \code{Tyrol}, \code{Upper Austria},
\code{Vienna} and \code{Vorarlberg}).}
\item{age}{integer; the
person's age.}
\item{rb090}{factor; the person's gender (levels
\code{male} and \code{female}).}
\item{pid}{personal ID}
\item{weight}{sampling weights}
\item{pl031}{factor; the person's
economic status (levels \code{1} = working full time, \code{2} = working
part time, \code{3} = unemployed, \code{4} = pupil, student, further
training or unpaid work experience or in compulsory military or community
service, \code{5} = in retirement or early retirement or has given up
business, \code{6} = permanently disabled or/and unfit to work or other
inactive person, \code{7} = fulfilling domestic tasks and care
responsibilities).}
\item{pb220a}{factor; the person's citizenship
(levels \code{AT}, \code{EU} and \code{Other}).}
\item{...}{see Eurostat's code book for further variables}
}}
\source{
This is a synthetic data set based on Austrian EU-SILC data from
2013.  The original sample was provided by Statistics Austria.
}
\description{
This data set is synthetically generated from real Austrian EU-SILC
(European Union Statistics on Income and Living Conditions) data 2013.
}
\details{
The data set consists of 5977 households and is used as sample data in some
of the examples in package \code{simPop}.  Note that it is included
for illustrative purposes only.  The sample weights do not reflect the true
population sizes of Austria and its regions.

62 variables of the original survey are
simulated for this example data set.  The variable names are rather cryptic
codes, but these are the standardized names used by the statistical
agencies.  Furthermore, the variables \code{hsize}, \code{age} and
\code{netIncome} are not included in the standardized format of EU-SILC
data, but have been derived from other variables for convenience.
}
\examples{
data(eusilc13puf)
str(eusilc13puf)
}
\author{
Matthias Templ
}
\references{
Eurostat (2013) Description of target variables: Cross-sectional
and longitudinal.
}
\keyword{datasets}

