% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{geometric_mean}
\alias{geometric_mean}
\alias{geometric_mean_reference}
\alias{geometric_mean}
\alias{arithmetic_mean}
\title{Geometric mean by factor}
\usage{
geometric_mean_reference(w, classes)

geometric_mean(w, classes)

arithmetic_mean(w, classes)
}
\arguments{
\item{w}{An numeric vector. All entries should be positive.}

\item{classes}{A factor variable. Must have the same length as \code{w}.}
}
\description{
This function calculates the geometric mean of the weight for each class. \code{geometric_mean} returns a \code{numeric} of
the same length as \code{w} which stores the averaged weight for each observation. \code{geometric_mean_reference} returns
the same value by reference, i.e. the input value \code{w} gets overwritten by the updated weights. See examples.
}
\examples{

## create random data
nobs <- 10
classLabels <- letters[1:3]
dat = data.frame(
  weight = exp(rnorm(nobs)),
  household = factor(sample(classLabels, nobs, replace = TRUE))
)
dat

## calculate weights with geometric_mean
new_weight <- geometric_mean(dat$weight, dat$household)
cbind(dat, new_weight)

## calculate weights "by reference"
geometric_mean_reference(dat$weight, dat$household)
dat

}
