% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getExclude}
\alias{getExclude}
\title{Position of missing values in data}
\usage{
getExclude(x, ...)
}
\arguments{
\item{x}{a vector, matrix, data.frame or data.table}

\item{...}{other arguments, not currently used}
}
\value{
Interger vector with positions indicating missing values
}
\description{
Get the positions of missing values in the data. This function is used internally
by other other functions of this package
}
