% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu2.r
\name{kishFactor}
\alias{kishFactor}
\title{Kish Factor}
\usage{
kishFactor(w)
}
\arguments{
\item{w}{a numeric vector with weights}
}
\value{
The function will return the the kish factor
}
\description{
Compute the kish factor for a specific weight vector
}
\examples{
kishFactor(rep(1,10))
kishFactor(rlnorm(10))
}
\author{
Alexander Kowarik
}
