% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0classes.R, R/dataSets.R
\docType{methods}
\name{simPopMethods}
\alias{popData}
\alias{popData,simPopObj-method}
\alias{popObj}
\alias{popObj,simPopObj-method}
\alias{popObj-methods}
\alias{popObj<-}
\alias{popObj<-,simPopObj,dataObj-method}
\alias{popObj<--methods}
\alias{sampleData}
\alias{sampleData,simPopObj-method}
\alias{sampleObj}
\alias{sampleObj,simPopObj-method}
\alias{sampleObj-methods}
\alias{sampleObj<-}
\alias{sampleObj<-,simPopObj,dataObj-method}
\alias{sampleObj<--methods}
\alias{simPopMethods}
\alias{tableObj}
\alias{tableObj,simPopObj-method}
\alias{tableObj-methods}
\title{Methods to query/set slots from objects of class
\code{\linkS4class{simPopObj}}.}
\description{
The methods allow to query or replace various information of
\code{\linkS4class{simPopObj}}-objects.

The methods allow to query or replace various information of
\code{\linkS4class{simPopObj}}-objects.
}
\section{Methods}{
 \describe{ Below, the methods to access/set various slots
of \code{\linkS4class{simPopObj}}-objects are shown.
\item{list("sampleObj")}{ returns slot 'sample' of the input object. }\item{
with }{ returns slot 'sample' of the input object.
}\item{list("signature(object=\"simPopObj\")")}{ returns slot 'sample' of
the input object. } \item{list("sampleObj<-")}{ replaces slot 'sample' of
the input object with the object \code{value}. }\item{ with }{ replaces slot
'sample' of the input object with the object \code{value}.
}\item{list("signature(object=\"simPopObj\", value=\"dataObj\")")}{ replaces
slot 'sample' of the input object with the object \code{value}. }
\item{list("sampleData")}{ returns slot 'data' of slot 'sample' of the input
object. }\item{ with }{ returns slot 'data' of slot 'sample' of the input
object. }\item{list("signature(object=\"simPopObj\")")}{ returns slot 'data'
of slot 'sample' of the input object. } \item{list("popObj")}{ returns slot
'pop' of the input object. }\item{ with }{ returns slot 'pop' of the input
object. }\item{list("signature(object=\"simPopObj\")")}{ returns slot 'pop'
of the input object. } \item{list("popObj<-")}{ replaces slot 'pop' of the
input object with the object \code{value}. }\item{ with }{ replaces slot
'pop' of the input object with the object \code{value}.
}\item{list("signature(object=\"simPopObj\", value=\"dataObj\")")}{ replaces
slot 'pop' of the input object with the object \code{value}. }
\item{list("popData")}{ returns slot 'data' of slot 'pop' of the input
object. }\item{ with }{ returns slot 'data' of slot 'pop' of the input
object. }\item{list("signature(object=\"simPopObj\")")}{ returns slot 'data'
of slot 'pop' of the input object. } \item{list("tableObj")}{ returns slot
'table' of the input object. }\item{ with }{ returns slot 'table' of the
input object. }\item{list("signature(object=\"simPopObj\")")}{ returns slot
'table' of the input object. } }
}

\section{Methods}{
 \describe{ Below, the methods to access/set various slots
of \code{\linkS4class{simPopObj}}-objects are shown.
\item{list("sampleObj")}{ returns slot 'sample' of the input object. }\item{
with }{ returns slot 'sample' of the input object.
}\item{list("signature(object=\"simPopObj\")")}{ returns slot 'sample' of
the input object. } \item{list("sampleObj<-")}{ replaces slot 'sample' of
the input object with the object \code{value}. }\item{ with }{ replaces slot
'sample' of the input object with the object \code{value}.
}\item{list("signature(object=\"simPopObj\", value=\"dataObj\")")}{ replaces
slot 'sample' of the input object with the object \code{value}. }
\item{list("sampleData")}{ returns slot 'data' of slot 'sample' of the input
object. }\item{ with }{ returns slot 'data' of slot 'sample' of the input
object. }\item{list("signature(object=\"simPopObj\")")}{ returns slot 'data'
of slot 'sample' of the input object. } \item{list("popObj")}{ returns slot
'pop' of the input object. }\item{ with }{ returns slot 'pop' of the input
object. }\item{list("signature(object=\"simPopObj\")")}{ returns slot 'pop'
of the input object. } \item{list("popObj<-")}{ replaces slot 'pop' of the
input object with the object \code{value}. }\item{ with }{ replaces slot
'pop' of the input object with the object \code{value}.
}\item{list("signature(object=\"simPopObj\", value=\"dataObj\")")}{ replaces
slot 'pop' of the input object with the object \code{value}. }
\item{list("popData")}{ returns slot 'data' of slot 'pop' of the input
object. }\item{ with }{ returns slot 'data' of slot 'pop' of the input
object. }\item{list("signature(object=\"simPopObj\")")}{ returns slot 'data'
of slot 'pop' of the input object. } \item{list("tableObj")}{ returns slot
'table' of the input object. }\item{ with }{ returns slot 'table' of the
input object. }\item{list("signature(object=\"simPopObj\")")}{ returns slot
'table' of the input object. } }
}
\examples{
data(eusilcS)
inp <- specifyInput(data=eusilcS, hhid="db030", hhsize="hsize", strata="db040", weight="db090")
inpObj <- simStructure(data=inp, method="direct", basicHHvars=c("age", "rb090"))
class(inpObj)
str(sampleObj(inpObj))

inp2 <- specifyInput(data=eusilcS, hhid="db030", hhsize="hsize", strata="pb220a", weight="db090")
sampleObj(inpObj) <- inp2
str(sampleObj(inpObj))
data(eusilcS)
inp <- specifyInput(data=eusilcS, hhid="db030", hhsize="hsize", strata="db040", weight="db090")
inpObj <- simStructure(data=inp, method="direct", basicHHvars=c("age", "rb090"))
class(inpObj)
str(sampleObj(inpObj))

inp2 <- specifyInput(data=eusilcS, hhid="db030", hhsize="hsize", strata="pb220a", weight="db090")
sampleObj(inpObj) <- inp2
str(sampleObj(inpObj))
}
\keyword{methods}

