\name{sprague}
\alias{sprague}
\title{
Sprague multipliers
}
\description{
Sprague multipliers
}
\usage{
sprague(x, table = FALSE, breaks = c(seq(0, 80, 5), 150))
}
\arguments{
  \item{x}{
numeric vector of ages
}
  \item{table}{
more later
}
  \item{breaks}{
more later
}
}
\details{
comming soon
}
\value{
coming soon
}
\references{
coming soon
}
\author{
Matthias Templ
}
\seealso{
\code{\link{whipple}}
}
\examples{
x <- sample(1:100, 5000, TRUE)
sprague(x)
}
\keyword{ arith }
