% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0classes.R
\docType{class}
\name{dataObj-class}
\alias{dataObj-class}
\alias{show,dataObj-method}
\title{Class \code{"dataObj"}}
\description{
Objects of this class contain information on a population or survey.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("dataObj", ...)} but are usually automatically created when using
\code{\link{simStructure}}.
}

\examples{
showClass("dataObj")

## show method, generate an object of class dataObj first
data(eusilcS)
inp <- specifyInput(data=eusilcS, hhid="db030", weight="rb050", strata="db040")
## shows some basic information:
inp

}
\seealso{
\code{\linkS4class{simPopObj}}
}
\author{
Bernhard Meindl and Matthias Templ
}
\keyword{classes}
