% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets.R
\docType{data}
\name{totalsRG}
\alias{totalsRG}
\alias{totalsRGtab}
\title{Population totals Region times Gender for Austria 2006}
\format{
totalsRG: A data frame with 18 observations on the following 3
variables.  \describe{ \item{list("rb090")}{gender; a factor with levels
\code{female} \code{male}} \item{list("db040")}{region; a factor with levels
\code{Burgenland} \code{Carinthia} \code{Lower Austria,} \code{Salzburg}
\code{Styria} \code{Tyrol} \code{Upper Austria} \code{Vienna}
\code{Vorarlberg}} \item{list("Freq")}{totals; a numeric vector} }
totalsRGtab: a two-dimensional table holding the same information

totalsRG: A data frame with 18 observations on the following 3
variables.  \describe{ \item{list("rb090")}{gender; a factor with levels
\code{female} \code{male}} \item{list("db040")}{region; a factor with levels
\code{Burgenland} \code{Carinthia} \code{Lower Austria,} \code{Salzburg}
\code{Styria} \code{Tyrol} \code{Upper Austria} \code{Vienna}
\code{Vorarlberg}} \item{list("Freq")}{totals; a numeric vector} }
totalsRGtab: a two-dimensional table holding the same information
}
\source{
StatCube - statistical data base,
\url{http://www.statistik.at}

StatCube - statistical data base,
\url{http://www.statistik.at/}
}
\description{
Population characteristics Region times Gender from Austria.

Using \code{\link{samp}} \code{\link{samp<-}} it is possible to extract or
rather modify variables of the sample data within slot \code{data} in slot
\code{sample} of the \code{\link{simPopObj-class}}-object. Using
\code{\link{pop}} \code{\link{pop<-}} it is possible to extract or rather
modify variables of the synthetic population within in slot \code{data} in
slot \code{sample} of the \code{\link{simPopObj-class}}-object.
}
\details{
Population totals Region times Gender for Austria 2006

Population characteristics Region times Gender from Austria.
}
\examples{
data(totalsRG)
totalsRG
data(totalsRGtab)
totalsRGtab
data(totalsRG)
totalsRG
data(totalsRGtab)
totalsRGtab
}
\keyword{datasets}
